+++ lib/binding/Auth/login_bindiing.dart
@@ -1,7 +1,7 @@
 import 'package:colab_ezzyfy_solutions/controller/Auth/login_controller.dart';
 import 'package:get/get.dart';
 
-import '../../controller/Auth/firebase_controller.dart';
+import '../firebase/firebase_controller.dart';
 import '../../controller/Auth/register_controller.dart';
