+++ lib/binding/home_dashboard_bindiing.dart
@@ -2,7 +2,7 @@ import 'package:colab_ezzyfy_solutions/controller/Auth/login_controller.dart';
 import 'package:colab_ezzyfy_solutions/controller/home_dashboard_controller.dart';
 import 'package:get/get.dart';
 
-import '../../controller/Auth/firebase_controller.dart';
+import 'firebase/firebase_controller.dart';
 
 
 class HomeDashboardBinding extends Bindings {
