+++ lib/binding/splash_binding.dart
@@ -1,8 +1,8 @@
-import 'package:colab_ezzyfy_solutions/controller/Auth/firebase_controller.dart';
+import 'package:colab_ezzyfy_solutions/binding/firebase/firebase_controller.dart';
 import 'package:colab_ezzyfy_solutions/controller/splash_controller.dart';
 import 'package:get/get.dart';
 
-import '../resource/firebase_dayabase_schema.dart';
+import '../resource/firebase_database_schema.dart';
