+++ lib/resource/firebase_database_schema.dart
@@ -1,5 +1,3 @@
-
-
 import 'package:get/get.dart';
 
 class FirebaseDatabaseSchema extends GetxController {
